+++ packages/mobile/App/models/LocalSystemFact.ts
@@ -1,7 +1,7 @@
 import { Entity, Column } from 'typeorm/browser';
 
 import { BaseModel } from './BaseModel';
-import { SYNC_DIRECTIONS } from './constants';
+import { SYNC_DIRECTIONS } from './types';
 
 @Entity('local_system_fact')
 export class LocalSystemFact extends BaseModel {
