--- packages/mobile/App/models/constants.ts
@@ -1,6 +1,0 @@
-export enum SYNC_DIRECTIONS {
-  DO_NOT_SYNC = 'do_not_sync',
-  PUSH_TO_CENTRAL = 'push_to_central',
-  PULL_FROM_CENTRAL = 'pull_from_central',
-  BIDIRECTIONAL = 'bidirectional',
-}
